+++ app/tests/unit/test_config.py
@@ -20,6 +20,8 @@ from github_runner_image_builder.config import (
     [
         pytest.param(Arch.ARM64, "aarch64", id="arm64"),
         pytest.param(Arch.X64, "x86_64", id="amd64"),
+        pytest.param(Arch.S390X, "s390x", id="s390x"),
+        pytest.param(Arch.PPC64LE, "ppc64le", id="ppc64le"),
     ],
 )
 def test_arch_openstack_conversion(arch: Arch, expected: str):
@@ -35,7 +37,7 @@ def test_arch_openstack_conversion(arch: Arch, expected: str):
     "image",
     [
         pytest.param("dingo", id="dingo"),
-        pytest.param("focal", id="focal"),
+        pytest.param("bionic", id="bionic"),
         pytest.param("firefox", id="firefox"),
     ],
 )
@@ -70,6 +72,7 @@ def test_base_image(image: str, expected_base_image: BaseImage):
 @pytest.mark.parametrize(
     "base_image, expected_version",
     [
+        pytest.param(BaseImage.FOCAL, "20.04", id="focal"),
         pytest.param(BaseImage.JAMMY, "22.04", id="jammy"),
         pytest.param(BaseImage.NOBLE, "24.04", id="noble"),
         pytest.param(None, None, id="None"),