+++ src/state.py
@@ -7,7 +7,6 @@ import dataclasses
 import logging
 import multiprocessing
 import os
-import platform
 import typing
 import urllib.parse
 from enum import Enum
@@ -18,9 +17,11 @@ import pydantic
 logger = logging.getLogger(__name__)
 
 ARCHITECTURES_ARM64 = {"aarch64", "arm64"}
-ARCHITECTURES_X86 = {"x86_64", "amd64"}
+ARCHITECTURES_S390x = {"s390x"}
+ARCHITECTURES_PPC64LE = {"ppc64le", "ppc64el"}
+ARCHITECTURES_X86 = {"x86_64", "amd64", "x64"}
 CLOUD_NAME = "builder"
-LTS_IMAGE_VERSION_TAG_MAP = {"22.04": "jammy", "24.04": "noble"}
+LTS_IMAGE_VERSION_TAG_MAP = {"20.04": "focal", "22.04": "jammy", "24.04": "noble"}
 
 ARCHITECTURE_CONFIG_NAME = "architecture"
 BASE_IMAGE_CONFIG_NAME = "base-image"
@@ -70,6 +71,8 @@ class Arch(str, Enum):
     Attributes:
         ARM64: Represents an ARM64 system architecture.
         X64: Represents an X64/AMD64 system architecture.
+        S390X: Represents an S390X system architecture.
+        PPC64LE: Represents an PPC64LE system architecture.
     """
 
     def __str__(self) -> str:
@@ -82,6 +85,8 @@ class Arch(str, Enum):
 
     ARM64 = "arm64"
     X64 = "x64"
+    S390X = "s390x"
+    PPC64LE = "ppc64le"
 
     @classmethod
     def from_charm(cls, charm: ops.CharmBase) -> "Arch":
@@ -89,6 +94,9 @@ class Arch(str, Enum):
 
         Args:
             charm: The charm instance.
+
+        Raises:
+            UnsupportedArchitectureError: if the configured architecture is unsupported.
 
         Returns:
             The architecture to build for.
@@ -101,31 +109,16 @@ class Arch(str, Enum):
                 return Arch.ARM64
             case arch if arch in ARCHITECTURES_X86:
                 return Arch.X64
+            case arch if arch in ARCHITECTURES_S390x:
+                return Arch.S390X
+            case arch if arch in ARCHITECTURES_PPC64LE:
+                return Arch.PPC64LE
             case _:
-                return _get_supported_arch()
+                raise UnsupportedArchitectureError(msg=f"Unsupported {arch=}")
 
 
 class UnsupportedArchitectureError(CharmConfigInvalidError):
     """Raised when given machine charm architecture is unsupported."""
-
-
-def _get_supported_arch() -> Arch:
-    """Get current machine architecture.
-
-    Raises:
-        UnsupportedArchitectureError: if the current architecture is unsupported.
-
-    Returns:
-        Arch: Current machine architecture.
-    """
-    arch = platform.machine()
-    match arch:
-        case arch if arch in ARCHITECTURES_ARM64:
-            return Arch.ARM64
-        case arch if arch in ARCHITECTURES_X86:
-            return Arch.X64
-        case _:
-            raise UnsupportedArchitectureError(msg=f"Unsupported {arch=}")
 
 
 class InvalidBaseImageError(CharmConfigInvalidError):
@@ -136,10 +129,12 @@ class BaseImage(str, Enum):
     """The ubuntu OS base image to build and deploy runners on.
 
     Attributes:
+        FOCAL: The focal ubuntu LTS image.
         JAMMY: The jammy ubuntu LTS image.
         NOBLE: The noble ubuntu LTS image.
     """
 
+    FOCAL = "focal"
     JAMMY = "jammy"
     NOBLE = "noble"
 
@@ -167,7 +162,7 @@ class BaseImage(str, Enum):
         image_names = tuple(
             image_name.lower().strip()
             for image_name in typing.cast(
-                str, charm.config.get(BASE_IMAGE_CONFIG_NAME, "jammy")
+                str, charm.config.get(BASE_IMAGE_CONFIG_NAME, "noble")
             ).split(",")
         )
         try: